WIP - first iteration of code update to version 3.0; - used compact.sqlite3 database for version 1.2; - login server is working; - you can enter the lobby; - you can create characters for version 1.2; - you can reconnect the server; - entering the world does not work, you need to fix packages.

+++ AAEmu.Game/Models/Game/Char/CharacterSkills.cs
@@ -37,9 +37,9 @@ namespace AAEmu.Game.Models.Game.Char
         {
             var template = SkillManager.Instance.GetSkillTemplate(skillId);
             if (template.AbilityId > 0 &&
-                template.AbilityId != (byte) Owner.Ability1 &&
-                template.AbilityId != (byte) Owner.Ability2 &&
-                template.AbilityId != (byte) Owner.Ability3)
+                template.AbilityId != (int) Owner.Ability1 &&
+                template.AbilityId != (int) Owner.Ability2 &&
+                template.AbilityId != (int) Owner.Ability3)
                 return;
             var points = ExpirienceManager.Instance.GetSkillPointsForLevel(Owner.Level);
             points -= GetUsedSkillPoints();
@@ -70,9 +70,9 @@ namespace AAEmu.Game.Models.Game.Char
         {
             var template = SkillManager.Instance.GetPassiveBuffTemplate(buffId);
             if(template.AbilityId > 0 && 
-               template.AbilityId != (byte)Owner.Ability1 && 
-               template.AbilityId != (byte)Owner.Ability2 && 
-               template.AbilityId != (byte)Owner.Ability3)
+               template.AbilityId != (int)Owner.Ability1 && 
+               template.AbilityId != (int)Owner.Ability2 && 
+               template.AbilityId != (int)Owner.Ability3)
                 return;
             var points = ExpirienceManager.Instance.GetSkillPointsForLevel(Owner.Level);
             points -= GetUsedSkillPoints();
